backend/requirements.txt
fastapi
uvicorn
python-jobspy
groq
python-dotenv